chore: rename hook and methods
